Expose code to partial classes
