+++ Equality/ViewModels/Boards/BoardPageViewModel.cs
@@ -12,8 +12,6 @@ using Catel.ExceptionHandling;
 
 using Equality.Data;
 using Equality.Extensions;
-using Equality.Helpers;
-using Equality.Http;
 using Equality.Models;
 using Equality.MVVM;
 using Equality.Services;
@@ -21,7 +19,7 @@ using Equality.Validation;
 
 namespace Equality.ViewModels
 {
-    public class BoardPageViewModel : ViewModel
+    public partial class BoardPageViewModel : ViewModel
     {
         #region DesignModeConstructor
 
@@ -62,32 +60,13 @@ namespace Equality.ViewModels
 
         protected INavigationService NavigationService { get; set; }
 
-        protected IColumnService ColumnService { get; set; }
-
-        protected ICardService CardService { get; set; }
-
         public BoardPageViewModel(INavigationService navigationService, IColumnService columnService, ICardService cardService)
         {
             NavigationService = navigationService;
-            ColumnService = columnService;
-            CardService = cardService;
+            RegisterColumnCommands(columnService);
+            RegisterCardCommands(cardService);
 
             ToBoards = new(OnToBoardsExecute);
-            OpenCreateColumnWindow = new(OnOpenCreateColumnWindowExecuteAsync);
-            OpenCreateCardWindow = new(OnOpenCreateCardWindowExecuteAsync);
-
-            StartEditColumn = new(OnStartEditColumnExecuteAsync);
-            SaveNewColumnName = new(OnSaveNewColumnNameExecuteAsync, () => GetFieldErrors(nameof(NewColumnName)) == string.Empty);
-            CancelEditColumn = new(OnCancelEditColumnExecute);
-            UpdateColumnOrder = new(OnUpdateColumnOrderExecuteAsync);
-            DeleteColumn = new(OnDeleteColumnExecuteAsync);
-
-            StartEditCard = new(OnStartEditCardExecuteAsync);
-            SaveNewCardName = new(OnSaveNewCardNameExecuteAsync, () => GetFieldErrors(nameof(NewCardName)) == string.Empty);
-            CancelEditCard = new(OnCancelEditCardExecute);
-            UpdateCardOrder = new(OnUpdateCardOrderExecuteAsync);
-            DeleteCard = new(OnDeleteCardExecuteAsync);
-            MoveCardToColumn = new(OnMoveCardToColumnAsync);
         }
 
         #region Properties
@@ -97,39 +76,6 @@ namespace Equality.ViewModels
         public ObservableCollection<Column> Columns { get; set; } = new();
 
         public bool IsDragging => DragCard != null || DragColumn != null;
-
-        #region ColumnProperties
-
-        public Column DragColumn { get; set; }
-
-        public CreateColumnControlViewModel CreateColumnVm { get; set; }
-
-        public Column EditableColumn { get; set; }
-
-        [Validatable]
-        public string NewColumnName { get; set; }
-
-        #endregion ColumnProperties
-
-        #region CardProperties
-
-        public Card DragCard { get; set; }
-
-        /// <summary>
-        /// The column in which the draggable card.
-        /// </summary>
-        public Column DraggableCardColumn { get; set; }
-
-        public CreateCardControlViewModel CreateCardVm { get; set; }
-
-        public Column ColumnForNewCard { get; set; }
-
-        public Card EditableCard { get; set; }
-
-        [Validatable]
-        public string NewCardName { get; set; }
-
-        #endregion CardProperties
 
         #endregion
 
@@ -142,301 +88,9 @@ namespace Equality.ViewModels
             NavigationService.Navigate<BoardsPageViewModel, ProjectPageViewModel>();
         }
 
-        #region CreateColumn
-
-        public TaskCommand OpenCreateColumnWindow { get; private set; }
-
-        private async Task OnOpenCreateColumnWindowExecuteAsync()
-        {
-            CreateColumnVm = MvvmHelper.CreateViewModel<CreateColumnControlViewModel>();
-            CreateColumnVm.ClosedAsync += CreateColumnVmClosedAsync;
-        }
-
-        private Task CreateColumnVmClosedAsync(object sender, ViewModelClosedEventArgs e)
-        {
-            CreateColumnVm.ClosedAsync -= CreateColumnVmClosedAsync;
-
-            if (CreateColumnVm.Result) {
-                Columns.Add(CreateColumnVm.Column);
-
-                // Open control again.
-                OpenCreateColumnWindow.Execute();
-            } else {
-                CreateColumnVm = null;
-            }
-
-            return Task.CompletedTask;
-        }
-
-        #endregion CreateColumn
-
-        #region EditColumn
-
-        public Command<Column> StartEditColumn { get; private set; }
-
-        private void OnStartEditColumnExecuteAsync(Column column)
-        {
-            NewColumnName = column.Name;
-            EditableColumn = column;
-        }
-
-        public Command CancelEditColumn { get; private set; }
-
-        private void OnCancelEditColumnExecute()
-        {
-            EditableColumn = null;
-            NewColumnName = null;
-            Validate(true);
-        }
-
-        public TaskCommand SaveNewColumnName { get; private set; }
-
-        private async Task OnSaveNewColumnNameExecuteAsync()
-        {
-            if (FirstValidationHasErrors()) {
-                return;
-            }
-
-            if (NewColumnName == EditableColumn.Name) {
-                CancelEditColumn.Execute();
-
-                return;
-            }
-
-            try {
-                Column column = new()
-                {
-                    Id = EditableColumn.Id,
-                    Name = NewColumnName,
-                };
-
-                var response = await ColumnService.UpdateColumnAsync(column);
-                EditableColumn.Name = response.Object.Name;
-
-                CancelEditColumn.Execute();
-            } catch (UnprocessableEntityHttpException e) {
-                HandleApiErrors(e.Errors, new() { { "name", nameof(NewColumnName) } });
-            } catch (HttpRequestException e) {
-                Data.ExceptionHandler.Handle(e);
-            }
-        }
-
-        #endregion EditColumn
-
-        #region UpdateColumnOrder
-
-        public TaskCommand UpdateColumnOrder { get; private set; }
-
-        private async Task OnUpdateColumnOrderExecuteAsync()
-        {
-            if (DragColumn == null) {
-                return;
-            }
-
-            try {
-                var afterColumn = Columns
-                    .TakeWhile(col => col.Id != DragColumn.Id)
-                    .LastOrDefault();
-
-                await ColumnService.UpdateColumnOrderAsync(DragColumn, afterColumn);
-            } catch (HttpRequestException e) {
-                Data.ExceptionHandler.Handle(e);
-            }
-        }
-
-        #endregion UpdateColumnOrder
-
-        #region DeleteColumn
-
-        public TaskCommand<Column> DeleteColumn { get; private set; }
-
-        private async Task OnDeleteColumnExecuteAsync(Column column)
-        {
-            var view = new Views.DeleteColumnDialog();
-            bool result = (bool)await MaterialDesignThemes.Wpf.DialogHost.Show(view);
-            if (!result) {
-                return;
-            }
-
-            try {
-                await ColumnService.DeleteColumnAsync(column);
-
-                Columns.Remove(column);
-            } catch (HttpRequestException e) {
-                Data.ExceptionHandler.Handle(e);
-            }
-        }
-
-        #endregion DeleteColumn
-
-        #region CreateCard
-
-        public TaskCommand<Column> OpenCreateCardWindow { get; private set; }
-
-        private async Task OnOpenCreateCardWindowExecuteAsync(Column column)
-        {
-            if (CreateCardVm != null) {
-                CreateCardVm.ClosedAsync -= CreateCardVmClosedAsync;
-            }
-
-            ColumnForNewCard = column;
-
-            CreateCardVm = MvvmHelper.CreateViewModel<CreateCardControlViewModel>(ColumnForNewCard);
-            CreateCardVm.ClosedAsync += CreateCardVmClosedAsync;
-        }
-
-        private Task CreateCardVmClosedAsync(object sender, ViewModelClosedEventArgs e)
-        {
-            CreateCardVm.ClosedAsync -= CreateCardVmClosedAsync;
-
-            if (CreateCardVm.Result) {
-                ColumnForNewCard.Cards.Add(CreateCardVm.Card);
-
-                // Open control again.
-                OpenCreateCardWindow.Execute(ColumnForNewCard);
-            } else {
-                CreateCardVm = null;
-                ColumnForNewCard = null;
-            }
-
-            return Task.CompletedTask;
-        }
-
-        #endregion CreateCard
-
-        #region EditCard
-
-        public Command<Card> StartEditCard { get; private set; }
-
-        private void OnStartEditCardExecuteAsync(Card card)
-        {
-            NewCardName = card.Name;
-            EditableCard = card;
-        }
-
-        public Command CancelEditCard { get; private set; }
-
-        private void OnCancelEditCardExecute()
-        {
-            EditableCard = null;
-            NewCardName = null;
-            Validate(true);
-        }
-
-        public TaskCommand SaveNewCardName { get; private set; }
-
-        private async Task OnSaveNewCardNameExecuteAsync()
-        {
-            if (FirstValidationHasErrors()) {
-                return;
-            }
-
-            if (NewCardName == EditableCard.Name) {
-                CancelEditCard.Execute();
-
-                return;
-            }
-
-            try {
-                Card card = new()
-                {
-                    Id = EditableCard.Id,
-                    Name = NewCardName,
-                };
-
-                var response = await CardService.UpdateCardAsync(card);
-
-                EditableCard.SyncWith(response.Object);
-
-                CancelEditCard.Execute();
-            } catch (UnprocessableEntityHttpException e) {
-                HandleApiErrors(e.Errors, new() { { "name", nameof(NewCardName) } });
-            } catch (HttpRequestException e) {
-                Data.ExceptionHandler.Handle(e);
-            }
-        }
-
-        #endregion EditCard
-
-        #region UpdateCardOrder
-
-        public TaskCommand UpdateCardOrder { get; private set; }
-
-        private async Task OnUpdateCardOrderExecuteAsync()
-        {
-            if (DragCard == null) {
-                return;
-            }
-
-            try {
-                var afterCard = DraggableCardColumn.Cards
-                   .TakeWhile(card => card.Id != DragCard.Id)
-                   .LastOrDefault();
-
-                await CardService.UpdateCardOrderAsync(DragCard, afterCard);
-            } catch (HttpRequestException e) {
-                Data.ExceptionHandler.Handle(e);
-            }
-        }
-
-        #endregion UpdateCardOrder
-
-        #region MoveCard
-
-        public TaskCommand MoveCardToColumn { get; private set; }
-
-        private async Task OnMoveCardToColumnAsync()
-        {
-            if (DragCard == null || DraggableCardColumn == null) {
-                return;
-            }
-            try {
-                var afterCard = DraggableCardColumn.Cards
-                   .TakeWhile(card => card.Id != DragCard.Id)
-                   .LastOrDefault();
-
-                await CardService.MoveCardToColumnAsync(DragCard, DraggableCardColumn, afterCard);
-            } catch (HttpRequestException e) {
-                Data.ExceptionHandler.Handle(e);
-            }
-        }
-
-        #endregion MoveCard
-
-        #region DeleteCard
-
-        public TaskCommand<Card> DeleteCard { get; private set; }
-
-        private async Task OnDeleteCardExecuteAsync(Card card)
-        {
-            var view = new Views.DeleteCardDialog();
-            bool result = (bool)await MaterialDesignThemes.Wpf.DialogHost.Show(view);
-            if (!result) {
-                return;
-            }
-
-            try {
-                await CardService.DeleteCardAsync(card);
-
-                foreach (var column in Columns) {
-                    column.Cards.Remove(card);
-                }
-            } catch (HttpRequestException e) {
-                Data.ExceptionHandler.Handle(e);
-            }
-        }
-
-        #endregion DeleteCard
-
         #endregion
 
         #region Methods
-
-        public void MoveCard(Card card, Column oldColumn, Column newColumn, int index)
-        {
-            oldColumn.Cards.Remove(card);
-            newColumn.Cards.Insert(index, card);
-        }
 
         public (Column, Card) FindColumnAndCard(ulong cardId)
         {
@@ -474,222 +128,14 @@ namespace Equality.ViewModels
 
         protected async Task SubscribePusherAsync()
         {
-            // Columns.
-            await ColumnService.SubscribeCreateColumnAsync(StateManager.SelectedBoard, (Column col, ulong? afterColumnId) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    if (afterColumnId == null) {
-                        Columns.Add(col);
-
-                        return;
-                    }
-
-                    if (afterColumnId == 0) {
-                        Columns.Insert(0, col);
-
-                        return;
-                    }
-
-                    var afterColumn = Columns.FirstOrDefault(col => col.Id == afterColumnId);
-                    if (afterColumn == null) {
-                        Columns.Add(col);
-                    } else {
-                        Columns.Insert(Columns.IndexOf(afterColumn) + 1, col);
-                    }
-                });
-            });
-            await ColumnService.SubscribeUpdateColumnAsync(StateManager.SelectedBoard, (Column column) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    var col = Columns.FirstOrDefault(col => col.Id == column.Id);
-                    if (col != null) {
-                        col.SyncWithOnly(column, new string[]
-                        {
-                            nameof(col.Name),
-                            nameof(col.CreatedAt),
-                            nameof(col.UpdatedAt),
-                        });
-                    }
-                });
-            });
-            await ColumnService.SubscribeUpdateColumnOrderAsync(StateManager.SelectedBoard, (ulong columnId, ulong afterId) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    var currCol = Columns.FirstOrDefault(col => col.Id == columnId);
-                    if (currCol == null) {
-                        return;
-                    }
-
-                    if (afterId == 0) {
-                        Columns.Remove(currCol);
-                        Columns.Insert(0, currCol);
-
-                        return;
-                    }
-
-                    var afterCol = Columns.FirstOrDefault(col => col.Id == afterId);
-                    if (afterCol != null) {
-                        Columns.Remove(currCol);
-                        Columns.Insert(Columns.IndexOf(afterCol) + 1, currCol);
-                    }
-                });
-            });
-            await ColumnService.SubscribeDeleteColumnAsync(StateManager.SelectedBoard, (ulong columnId) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    var col = Columns.FirstOrDefault(col => col.Id == columnId);
-                    if (col == null) {
-                        return;
-                    }
-
-                    // Disable dragging mode for the column.
-                    if (DragColumn.Id == columnId) {
-                        DragColumn = null;
-                    }
-
-                    Columns.Remove(col);
-                });
-            });
-
-            // Cards.
-            await CardService.SubscribeCreateCardAsync(StateManager.SelectedBoard, (Card card, ulong columnId, ulong? afterCardId) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    var column = Columns.FirstOrDefault(col => col.Id == columnId);
-                    if (column == null) {
-                        return;
-                    }
-
-                    if (afterCardId == null) {
-                        column.Cards.Add(card);
-
-                        return;
-                    }
-
-                    if (afterCardId == 0) {
-                        column.Cards.Insert(0, card);
-
-                        return;
-                    }
-
-                    var afterColumn = column.Cards.FirstOrDefault(col => col.Id == afterCardId);
-                    if (afterColumn == null) {
-                        column.Cards.Add(card);
-                    } else {
-                        column.Cards.Insert(column.Cards.IndexOf(afterColumn) + 1, card);
-                    }
-                });
-            });
-            await CardService.SubscribeUpdateCardAsync(StateManager.SelectedBoard, (Card updatedCard) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    var (_, card) = FindColumnAndCard(updatedCard.Id);
-
-                    if (card != null) {
-                        card.SyncWithOnly(updatedCard, new string[]
-                        {
-                            nameof(updatedCard.Name),
-                            nameof(updatedCard.CreatedAt),
-                            nameof(updatedCard.UpdatedAt),
-                        });
-                    }
-                });
-            });
-            await CardService.SubscribeUpdateCardOrderAsync(StateManager.SelectedBoard, (ulong cardId, ulong afterId) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    var (column, card) = FindColumnAndCard(cardId);
-                    if (column == null || card == null) {
-                        return;
-                    }
-
-                    if (afterId == 0) {
-                        column.Cards.Move(column.Cards.IndexOf(card), 0);
-
-                        return;
-                    }
-
-                    var afterCard = column.Cards.FirstOrDefault(card => card.Id == afterId);
-                    if (afterCard != null) {
-                        // If we move card to up, then we need to +1 to our index,
-                        // so we insert new card after the specified one, not before.
-                        int cardIndex = column.Cards.IndexOf(card);
-                        int afterCardIndex = column.Cards.IndexOf(afterCard);
-                        if (cardIndex > afterCardIndex) {
-                            afterCardIndex++;
-                        }
-
-                        column.Cards.Move(cardIndex, afterCardIndex);
-                    }
-                });
-            });
-            await CardService.SubscribeMoveCardToColumnAsync(StateManager.SelectedBoard, (ulong cardId, ulong columnId, ulong afterId) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    var newColumn = Columns.FirstOrDefault(c => c.Id == columnId);
-                    if (newColumn == null) {
-                        return;
-                    }
-
-                    var (oldColumn, card) = FindColumnAndCard(cardId);
-                    if (oldColumn == null || card == null) {
-                        return;
-                    }
-
-                    if (afterId == 0) {
-                        MoveCard(card, oldColumn, newColumn, 0);
-
-                        return;
-                    }
-
-                    var afterCard = newColumn.Cards.FirstOrDefault(card => card.Id == afterId);
-                    if (afterCard != null) {
-                        MoveCard(card, oldColumn, newColumn, newColumn.Cards.IndexOf(afterCard) + 1);
-                    }
-                });
-            });
-            await CardService.SubscribeDeleteCardAsync(StateManager.SelectedBoard, (ulong cardId) =>
-            {
-                App.Current.Dispatcher.Invoke(() =>
-                {
-                    var (column, card) = FindColumnAndCard(cardId);
-                    if (column == null || card == null) {
-                        return;
-                    }
-
-                    // Disable dragging mode for the card.
-                    if (DragCard.Id == cardId) {
-                        DragCard = null;
-                        DraggableCardColumn = null;
-                    }
-
-                    column.Cards.Remove(card);
-                });
-            });
+            RegisterPusherForColumns();
+            RegisterPusherForCards();
         }
 
         protected void UnsubscribePusherAsync()
         {
-            // Columns.
-            ColumnService.UnsubscribeCreateColumn(StateManager.SelectedBoard);
-            ColumnService.UnsubscribeUpdateColumn(StateManager.SelectedBoard);
-            ColumnService.UnsubscribeUpdateColumnOrder(StateManager.SelectedBoard);
-            ColumnService.UnsubscribeDeleteColumn(StateManager.SelectedBoard);
-
-            // Cards.
-            CardService.UnsubscribeCreateCard(StateManager.SelectedBoard);
-            CardService.UnsubscribeUpdateCard(StateManager.SelectedBoard);
-            CardService.UnsubscribeUpdateCardOrder(StateManager.SelectedBoard);
-            CardService.UnsubscribeMoveCardToColumn(StateManager.SelectedBoard);
-            CardService.UnsubscribeDeleteCard(StateManager.SelectedBoard);
+            UnregisterPusherForColumns();
+            UnregisterPusherForCards();
         }
 
         #endregion
